+++ test/commands/force/data/soql/query/dataSoqlQueryCommand.test.ts
@@ -34,7 +34,7 @@ describe('Execute a SOQL statement', function (): void {
       beforeEach(() => {
         soqlQuerySpy = sandbox
           .stub(SoqlQuery.prototype, 'runSoqlQuery')
-          .callsFake(() => Promise.resolve(soqlQueryExemplars.emptyQuery.soqlQueryResult));
+          .resolves(soqlQueryExemplars.emptyQuery.soqlQueryResult);
       });
       afterEach(() => {
         sandbox.restore();
@@ -111,7 +111,7 @@ describe('Execute a SOQL statement', function (): void {
       beforeEach(() => {
         soqlQuerySpy = sandbox
           .stub(SoqlQuery.prototype, 'runSoqlQuery')
-          .callsFake(() => Promise.resolve(soqlQueryExemplars.complexSubQuery.soqlQueryResult));
+          .resolves(soqlQueryExemplars.complexSubQuery.soqlQueryResult);
       });
 
       afterEach(() => {
@@ -148,11 +148,28 @@ describe('Execute a SOQL statement', function (): void {
           expect(ctx.stdout).to.include('records retrieved: 1');
         });
     });
+    describe('flag validation between --query and --soqlqueryfile', () => {
+      test
+        .withOrg({ username: 'test@org.com' }, true)
+        .stderr()
+        .command([
+          QUERY_COMMAND,
+          '--targetusername',
+          'test@org.com',
+          '--query',
+          'SELECT Amount, Id, Name,StageName, CloseDate, (SELECT Id,  ListPrice, PriceBookEntry.UnitPrice, PricebookEntry.Name, PricebookEntry.Id, PricebookEntry.product2.Family FROM OpportunityLineItems) FROM Opportunity',
+          '--soqlqueryfile',
+          'soql.txt',
+        ])
+        .it('should throw an error when both query (inline/file query) flags are specified', (ctx) => {
+          expect(ctx.stderr).to.include('--soqlqueryfile= cannot also be provided when using --query=');
+        });
+    });
     describe('reporters produce the correct aggregate query', () => {
       beforeEach(() => {
         soqlQuerySpy = sandbox
           .stub(SoqlQuery.prototype, 'runSoqlQuery')
-          .callsFake(() => Promise.resolve(soqlQueryExemplars.queryWithAgregates.soqlQueryResult));
+          .resolves(soqlQueryExemplars.queryWithAgregates.soqlQueryResult);
       });
       afterEach(() => {
         sandbox.restore();